hardware acceleration time! so many particles

+++ include/engine/types/positioning/Quad.hpp
@@ -2,7 +2,7 @@
 
 #include <SDL_rect.h>
 #include <vector>
-#include "EngineEnums.hpp"
+#include "../EngineEnums.hpp"
 
 namespace geng {
     /**
@@ -10,14 +10,14 @@ namespace geng {
     * @details You primary use the @code to_frame@endcode function to convert to a frame object. Frame objects can then take Vertex data and create render output from it.
     */
     struct Quad {
-        int x, y, w, h;
+        uint16_t x, y, w, h;
         float duration;
         GAnimType anim;
         /// Constructor for a quad. Can hold useful info for
         Quad(int x, int y, int w, int h,
             float duration = 0.f, GAnimType anim = GAnimType::IDLE);
 
-        /// Converts a quad into it's vertex points
+        /// Converts a quad into it's corresponding vertex points
         std::vector<SDL_FPoint> to_vert_points(int texW, int texH) const;
     };
 }
